+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+- Support multiple tracking events
+
 ## v.5.1.0 - 2020-08-12
 
 - Expose `validate` and `submit` in `Form` context to be able to create forms with manual validation and